requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0
greenlet==3.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AWS S3 integration for storage
boto3==1.28.62
botocore==1.31.63

# Image processing
Pillow==10.1.0

# Date handling
pytz==2023.3

# Caching and performance
redis==5.0.1
httpx==0.25.1

# Background tasks and queueing
celery==5.3.4
flower==2.0.1

# Testing
pytest>=8.3.3,<9.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.17.1

# Development tools
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1

qrcode[pil]>=7.3.1
Pillow>=9.5.0

pytest-env>=1.1.5